hack/util: Factor out some common code into a shared library
Simplify maintenance by giving us one place to edit this stuff.  Seems
useful enough to outweigh the downside, which is that the scripts now
need the sibling util.py in place and are no longer stand-alone.
